latitude/longitude: long to double

+++ src/main/java/fr/eql/ai116/duron/thomas/art/connect/entity/dto/SearchInputDto.java
@@ -4,8 +4,8 @@ import fr.eql.ai116.duron.thomas.art.connect.entity.ArtTag;
 import fr.eql.ai116.duron.thomas.art.connect.entity.EventType;
 
 public record SearchInputDto(
-        long latitude,
-        long longitude,
+        double latitude,
+        double longitude,
         String field,
         int perimeter,
         EventType eventType,
